requirements.txt
APScheduler==3.0.0
asgiref==3.5.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
black==22.12.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
daphne==4.0.0
dj-config-url==0.1.1
dj-database-url==1.0.0
Django==3.2.15
django-cors-headers==3.12.0
django-cors-middleware==1.5.0
django-crontab==0.7.1
django-model-utils==4.3.1
django-on-heroku==1.1.2
django-private-chat2==1.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==4.3.0
drf-yasg==1.21.4
firebase-admin==5.4.0
gevent==22.10.2
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
greenlet==2.0.1
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0
httplib2==0.21.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iso8601==1.1.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
mypy-extensions==0.4.3
numpy==1.24.1
packaging==21.3
pandas==1.5.3
pathspec==0.10.3
Pillow==9.2.0
platformdirs==2.6.0
proto-plus==1.22.2
protobuf==4.21.12
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfcm==1.5.4
PyJWT==1.7.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stripe==4.2.0
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
zope.event==4.6
zope.interface==5.5.2
